Partial-implementation of quality score
Don't have a camera to test with, but should show grey text widget when camera isn't connected, that turns red, yellow, or green depending on quality of focus quality score.

Working on changing text based on results, as well as adding this result to UI instead of its own window as it is currently.